task 1 done
This task is done with the csv file present in QGIS layer panel. In the assignment, it is not mentioned that where to read the csv file in the file explorer or the QGIS layers panels. If it is necessary to read the csv file from the file explorer then, let us know!